LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-not-application-expected.txt


PASS selectionStart on an input[type=hidden] returns null
PASS selectionEnd on an input[type=hidden] returns null
PASS selectionDirection on an input[type=hidden] returns null
PASS assigning selectionStart on an input[type=hidden] throws InvalidStateError
PASS assigning selectionEnd on an input[type=hidden] throws InvalidStateError
PASS assigning selectionDirection on an input[type=hidden] throws InvalidStateError
PASS setRangeText on an input[type=hidden] throws InvalidStateError
PASS setSelectionRange on an input[type=hidden] throws InvalidStateError
PASS selectionStart on an input[type=email] returns null
PASS selectionEnd on an input[type=email] returns null
PASS selectionDirection on an input[type=email] returns null
PASS assigning selectionStart on an input[type=email] throws InvalidStateError
PASS assigning selectionEnd on an input[type=email] throws InvalidStateError
PASS assigning selectionDirection on an input[type=email] throws InvalidStateError
PASS setRangeText on an input[type=email] throws InvalidStateError
PASS setSelectionRange on an input[type=email] throws InvalidStateError
FAIL selectionStart on an input[type=datetime-local] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionEnd on an input[type=datetime-local] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionDirection on an input[type=datetime-local] returns null assert_equals: expected (object) null but got (string) "none"
FAIL assigning selectionStart on an input[type=datetime-local] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionStart = 0;
      }" did not throw
FAIL assigning selectionEnd on an input[type=datetime-local] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionEnd = 0;
      }" did not throw
FAIL assigning selectionDirection on an input[type=datetime-local] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionDirection = 'none';
      }" did not throw
FAIL setRangeText on an input[type=datetime-local] throws InvalidStateError assert_throws_dom: function "function(){
        el.setRangeText("foobar");
      }" did not throw
FAIL setSelectionRange on an input[type=datetime-local] throws InvalidStateError assert_throws_dom: function "function(){
        el.setSelectionRange(0, 1);
      }" did not throw
FAIL selectionStart on an input[type=date] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionEnd on an input[type=date] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionDirection on an input[type=date] returns null assert_equals: expected (object) null but got (string) "none"
FAIL assigning selectionStart on an input[type=date] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionStart = 0;
      }" did not throw
FAIL assigning selectionEnd on an input[type=date] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionEnd = 0;
      }" did not throw
FAIL assigning selectionDirection on an input[type=date] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionDirection = 'none';
      }" did not throw
FAIL setRangeText on an input[type=date] throws InvalidStateError assert_throws_dom: function "function(){
        el.setRangeText("foobar");
      }" did not throw
FAIL setSelectionRange on an input[type=date] throws InvalidStateError assert_throws_dom: function "function(){
        el.setSelectionRange(0, 1);
      }" did not throw
FAIL selectionStart on an input[type=time] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionEnd on an input[type=time] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionDirection on an input[type=time] returns null assert_equals: expected (object) null but got (string) "none"
FAIL assigning selectionStart on an input[type=time] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionStart = 0;
      }" did not throw
FAIL assigning selectionEnd on an input[type=time] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionEnd = 0;
      }" did not throw
FAIL assigning selectionDirection on an input[type=time] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionDirection = 'none';
      }" did not throw
FAIL setRangeText on an input[type=time] throws InvalidStateError assert_throws_dom: function "function(){
        el.setRangeText("foobar");
      }" did not throw
FAIL setSelectionRange on an input[type=time] throws InvalidStateError assert_throws_dom: function "function(){
        el.setSelectionRange(0, 1);
      }" did not throw
PASS selectionStart on an input[type=number] returns null
PASS selectionEnd on an input[type=number] returns null
PASS selectionDirection on an input[type=number] returns null
PASS assigning selectionStart on an input[type=number] throws InvalidStateError
PASS assigning selectionEnd on an input[type=number] throws InvalidStateError
PASS assigning selectionDirection on an input[type=number] throws InvalidStateError
PASS setRangeText on an input[type=number] throws InvalidStateError
PASS setSelectionRange on an input[type=number] throws InvalidStateError
PASS selectionStart on an input[type=range] returns null
PASS selectionEnd on an input[type=range] returns null
PASS selectionDirection on an input[type=range] returns null
PASS assigning selectionStart on an input[type=range] throws InvalidStateError
PASS assigning selectionEnd on an input[type=range] throws InvalidStateError
PASS assigning selectionDirection on an input[type=range] throws InvalidStateError
PASS setRangeText on an input[type=range] throws InvalidStateError
PASS setSelectionRange on an input[type=range] throws InvalidStateError
FAIL selectionStart on an input[type=color] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionEnd on an input[type=color] returns null assert_equals: expected (object) null but got (number) 0
FAIL selectionDirection on an input[type=color] returns null assert_equals: expected (object) null but got (string) "none"
FAIL assigning selectionStart on an input[type=color] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionStart = 0;
      }" did not throw
FAIL assigning selectionEnd on an input[type=color] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionEnd = 0;
      }" did not throw
FAIL assigning selectionDirection on an input[type=color] throws InvalidStateError assert_throws_dom: function "function(){
        el.selectionDirection = 'none';
      }" did not throw
FAIL setRangeText on an input[type=color] throws InvalidStateError assert_throws_dom: function "function(){
        el.setRangeText("foobar");
      }" did not throw
FAIL setSelectionRange on an input[type=color] throws InvalidStateError assert_throws_dom: function "function(){
        el.setSelectionRange(0, 1);
      }" did not throw
PASS selectionStart on an input[type=checkbox] returns null
PASS selectionEnd on an input[type=checkbox] returns null
PASS selectionDirection on an input[type=checkbox] returns null
PASS assigning selectionStart on an input[type=checkbox] throws InvalidStateError
PASS assigning selectionEnd on an input[type=checkbox] throws InvalidStateError
PASS assigning selectionDirection on an input[type=checkbox] throws InvalidStateError
PASS setRangeText on an input[type=checkbox] throws InvalidStateError
PASS setSelectionRange on an input[type=checkbox] throws InvalidStateError
PASS selectionStart on an input[type=radio] returns null
PASS selectionEnd on an input[type=radio] returns null
PASS selectionDirection on an input[type=radio] returns null
PASS assigning selectionStart on an input[type=radio] throws InvalidStateError
PASS assigning selectionEnd on an input[type=radio] throws InvalidStateError
PASS assigning selectionDirection on an input[type=radio] throws InvalidStateError
PASS setRangeText on an input[type=radio] throws InvalidStateError
PASS setSelectionRange on an input[type=radio] throws InvalidStateError
PASS selectionStart on an input[type=file] returns null
PASS selectionEnd on an input[type=file] returns null
PASS selectionDirection on an input[type=file] returns null
PASS assigning selectionStart on an input[type=file] throws InvalidStateError
PASS assigning selectionEnd on an input[type=file] throws InvalidStateError
PASS assigning selectionDirection on an input[type=file] throws InvalidStateError
PASS setRangeText on an input[type=file] throws InvalidStateError
PASS setSelectionRange on an input[type=file] throws InvalidStateError
PASS selectionStart on an input[type=submit] returns null
PASS selectionEnd on an input[type=submit] returns null
PASS selectionDirection on an input[type=submit] returns null
PASS assigning selectionStart on an input[type=submit] throws InvalidStateError
PASS assigning selectionEnd on an input[type=submit] throws InvalidStateError
PASS assigning selectionDirection on an input[type=submit] throws InvalidStateError
PASS setRangeText on an input[type=submit] throws InvalidStateError
PASS setSelectionRange on an input[type=submit] throws InvalidStateError
PASS selectionStart on an input[type=image] returns null
PASS selectionEnd on an input[type=image] returns null
PASS selectionDirection on an input[type=image] returns null
PASS assigning selectionStart on an input[type=image] throws InvalidStateError
PASS assigning selectionEnd on an input[type=image] throws InvalidStateError
PASS assigning selectionDirection on an input[type=image] throws InvalidStateError
PASS setRangeText on an input[type=image] throws InvalidStateError
PASS setSelectionRange on an input[type=image] throws InvalidStateError
PASS selectionStart on an input[type=reset] returns null
PASS selectionEnd on an input[type=reset] returns null
PASS selectionDirection on an input[type=reset] returns null
PASS assigning selectionStart on an input[type=reset] throws InvalidStateError
PASS assigning selectionEnd on an input[type=reset] throws InvalidStateError
PASS assigning selectionDirection on an input[type=reset] throws InvalidStateError
PASS setRangeText on an input[type=reset] throws InvalidStateError
PASS setSelectionRange on an input[type=reset] throws InvalidStateError
PASS selectionStart on an input[type=button] returns null
PASS selectionEnd on an input[type=button] returns null
PASS selectionDirection on an input[type=button] returns null
PASS assigning selectionStart on an input[type=button] throws InvalidStateError
PASS assigning selectionEnd on an input[type=button] throws InvalidStateError
PASS assigning selectionDirection on an input[type=button] throws InvalidStateError
PASS setRangeText on an input[type=button] throws InvalidStateError
PASS setSelectionRange on an input[type=button] throws InvalidStateError
PASS selectionStart on an input[type=text] returns a value
PASS selectionEnd on an input[type=text] returns a value
PASS selectionDirection on an input[type=text] returns a value
PASS assigning selectionDirection "none" on an input[type=text] should give the initial direction
PASS assigning selectionStart on an input[type=text] doesn't throw an exception
PASS assigning selectionEnd on an input[type=text] doesn't throw an exception
PASS assigning selectionDirection on an input[type=text] doesn't throw an exception
PASS setRangeText on an input[type=text] doesn't throw an exception
PASS setSelectionRange on an input[type=text] doesn't throw an exception
PASS selectionStart on an input[type=search] returns a value
PASS selectionEnd on an input[type=search] returns a value
PASS selectionDirection on an input[type=search] returns a value
PASS assigning selectionDirection "none" on an input[type=search] should give the initial direction
PASS assigning selectionStart on an input[type=search] doesn't throw an exception
PASS assigning selectionEnd on an input[type=search] doesn't throw an exception
PASS assigning selectionDirection on an input[type=search] doesn't throw an exception
PASS setRangeText on an input[type=search] doesn't throw an exception
PASS setSelectionRange on an input[type=search] doesn't throw an exception
PASS selectionStart on an input[type=tel] returns a value
PASS selectionEnd on an input[type=tel] returns a value
PASS selectionDirection on an input[type=tel] returns a value
PASS assigning selectionDirection "none" on an input[type=tel] should give the initial direction
PASS assigning selectionStart on an input[type=tel] doesn't throw an exception
PASS assigning selectionEnd on an input[type=tel] doesn't throw an exception
PASS assigning selectionDirection on an input[type=tel] doesn't throw an exception
PASS setRangeText on an input[type=tel] doesn't throw an exception
PASS setSelectionRange on an input[type=tel] doesn't throw an exception
PASS selectionStart on an input[type=url] returns a value
PASS selectionEnd on an input[type=url] returns a value
PASS selectionDirection on an input[type=url] returns a value
PASS assigning selectionDirection "none" on an input[type=url] should give the initial direction
PASS assigning selectionStart on an input[type=url] doesn't throw an exception
PASS assigning selectionEnd on an input[type=url] doesn't throw an exception
PASS assigning selectionDirection on an input[type=url] doesn't throw an exception
PASS setRangeText on an input[type=url] doesn't throw an exception
PASS setSelectionRange on an input[type=url] doesn't throw an exception
PASS selectionStart on an input[type=password] returns a value
PASS selectionEnd on an input[type=password] returns a value
PASS selectionDirection on an input[type=password] returns a value
PASS assigning selectionDirection "none" on an input[type=password] should give the initial direction
PASS assigning selectionStart on an input[type=password] doesn't throw an exception
PASS assigning selectionEnd on an input[type=password] doesn't throw an exception
PASS assigning selectionDirection on an input[type=password] doesn't throw an exception
PASS setRangeText on an input[type=password] doesn't throw an exception
PASS setSelectionRange on an input[type=password] doesn't throw an exception
PASS selectionStart on an input[type=aninvalidtype] returns a value
PASS selectionEnd on an input[type=aninvalidtype] returns a value
PASS selectionDirection on an input[type=aninvalidtype] returns a value
PASS assigning selectionDirection "none" on an input[type=aninvalidtype] should give the initial direction
PASS assigning selectionStart on an input[type=aninvalidtype] doesn't throw an exception
PASS assigning selectionEnd on an input[type=aninvalidtype] doesn't throw an exception
PASS assigning selectionDirection on an input[type=aninvalidtype] doesn't throw an exception
PASS setRangeText on an input[type=aninvalidtype] doesn't throw an exception
PASS setSelectionRange on an input[type=aninvalidtype] doesn't throw an exception
PASS selectionStart on an input[type=null] returns a value
PASS selectionEnd on an input[type=null] returns a value
PASS selectionDirection on an input[type=null] returns a value
PASS assigning selectionDirection "none" on an input[type=null] should give the initial direction
PASS assigning selectionStart on an input[type=null] doesn't throw an exception
PASS assigning selectionEnd on an input[type=null] doesn't throw an exception
PASS assigning selectionDirection on an input[type=null] doesn't throw an exception
PASS setRangeText on an input[type=null] doesn't throw an exception
PASS setSelectionRange on an input[type=null] doesn't throw an exception
